requirements.txt
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pycodestyle==2.8.0
PyJWT==2.3.0
pymongo==4.0.2
python-dotenv==0.19.2
six==1.16.0
toml==0.10.2
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
